# backend/openapi.yaml
openapi: 3.0.3
info:
  title: CLISENSA API
  description: API REST de la plataforma multi-hospital de citas médicas CLISENSA.
  version: 0.5.0
servers:
  - url: http://localhost:4000/api
    description: Desarrollo local
...
paths:
...
  /citas:
    post:
      tags: [Citas]
      summary: Paciente reserva o médico asigna una cita directamente
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [medicoId, hospitalId, fecha, hora]
              properties:
                medicoId: { type: integer, example: 1 }
                fecha: { type: string, format: date, example: "2026-09-30" }
                hora: { type: string, example: "09:00" }
                hospitalId: { type: integer }
                servicioId: { type: integer }
                pacienteId: { type: integer, description: Identificador del paciente en asignaciones médicas }
                correo: { type: string, format: email }
                cedula: { type: string }
      responses:
        "201":
          description: Cita creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cita"
        "400": { description: Datos inválidos u horario inexistente }
        "401": { description: Falta autenticación }
        "403": { description: Se requiere rol paciente o médico de la agenda }
        "409": { description: Horario ya reservado }
    get:
      tags: [Citas]
      summary: Listar citas del paciente autenticado
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Citas del paciente
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: "#/components/schemas/Cita"
                    - type: object
                      properties:
                        medicoNombre: { type: string }
        "401": { description: Falta autenticación }
        "403": { description: Se requiere rol paciente }
  /citas/{id}/cancelar:
    put:
      tags: [Citas]
      summary: Cancelar cita propia (rol paciente)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Cita cancelada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cita"
        "401": { description: Falta autenticación }
        "403": { description: Cita ajena o rol incorrecto }
        "404": { description: Cita no encontrada }
  /citas/{id}/estado:
    put:
      tags: [Citas]
      summary: Actualizar estado de cita
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [estado]
              properties:
                estado: { type: string, enum: [CONFIRMADA, CANCELADA, ATENDIDA] }
      responses:
        "200": { description: Estado actualizado }
        "403": { description: No autorizado para esta cita }
  /citas/{id}/reprogramar:
    put:
      tags: [Citas]
      summary: Reprogramar cita propia (rol paciente)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fecha, hora]
              properties:
                fecha: { type: string, format: date, example: "2026-10-01" }
                hora: { type: string, example: "09:30" }
      responses:
        "200":
          description: Cita reprogramada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cita"
        "400": { description: Datos inválidos, cita cancelada u horario inexistente }
        "401": { description: Falta autenticación }
        "403": { description: Cita ajena o rol incorrecto }
        "404": { description: Cita no encontrada }
        "409": { description: Nuevo horario ya reservado }
  /citas/medico:
    get:
      tags: [Citas]
      summary: Consultar agenda del médico autenticado
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: fecha
          required: true
          schema: { type: string, format: date }
      responses:
        "200": { description: Agenda del médico }
        "401": { description: Falta autenticación }
        "403": { description: Se requiere rol médico }
  /citas/proximas:
    get:
      tags: [Citas]
      summary: Consultar las próximas ocho citas visibles para la cuenta autenticada
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Próximas citas activas }
        "401": { description: Falta autenticación }
        "403": { description: La cuenta no tiene hospital o agenda asociada }
  /citas/todas:
    get:
      tags: [Citas]
      summary: Consultar citas del hospital (personal) o de toda la red (ADMIN_SISTEMA)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: fecha
          required: false
          schema: { type: string, format: date }
      responses:
        "200": { description: Citas visibles para el rol autenticado }
        "401": { description: Falta autenticación }
        "403": { description: Se requiere rol administrativo }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Cita:
      type: object
      required: [id, pacienteId, medicoId, hospitalId, servicioId, fecha, hora, fechaHora, estado, creadaEn]
      properties:
        id: { type: integer }
        pacienteId: { type: integer }
        medicoId: { type: integer }
        hospitalId: { type: integer }
        servicioId: { type: integer }
        fechaHora: { type: string, format: date-time }
        fecha: { type: string, format: date }
        hora: { type: string, example: "09:00" }
        estado: { type: string, enum: [PENDIENTE, CONFIRMADA, CANCELADA, ATENDIDA] }
        creadaEn: { type: string, format: date-time }